backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
sqlparse==0.4.4
tzdata==2023.3
scapy==2.5.0
